requirements.txt
supabase
celery
redis
openai>=1.0.0
python-dotenv
reflex==0.7.13